third_party/blink/web_tests/svg/custom/svg-parse-overflow-2-expected.txt

CONSOLE ERROR: Error: <rect> attribute height: A negative value is not valid. ("-105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000")
